Define the architecture, set up Docker
The interfaces for exploring music with different backends are defined,
as well as the MusicDownloader to download music locally. The server is
listening for requests and handling them using the interfaces. Dockerfile
compiles the project, installs dependencies and runs all servers (Redis,
PostgreSQL, gRPC, the music browser itself), exposing only the music browser.